Monitor the cluster installation after configuring a BMH
This checks the remote cluster for an available ClusterVersion and for
all the nodes to be Ready. If these conditions are not met by the
timeout (default 1h) the cluster will be marked as failed.

The timeout is configurable using an annotation on the
ImageClusterInstall (`imageclusterinstall.extensions.hive.openshift.io/install-timeout`)

Resolves https://issues.redhat.com/browse/MGMT-17379